requirements.txt
websockets
python-dotenv
requests
pyyaml
orjson